fastApi_Stepik/pomodoro-time/pyproject.toml
[project]
name = "pomodoro-analytics"
version = "0.1.0"
description = ""
authors = [
    {name = "Anton Fayfer",email = "fayfer2004@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "black (>=25.1.0,<26.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "reddis (>=1.0.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
